+++ tests/testthat/test-nice_t_test.R
@@ -45,4 +45,13 @@ test_that("nice_t_test", {
     mu = 10,
     verbose = FALSE
   ))
+  
+  # Paired t-test instead of independent samples
+  expect_snapshot(nice_t_test(
+    data = ToothGrowth,
+    response = "len",
+    group = "supp",
+    paired = TRUE,
+    verbose = FALSE
+  ))
 })